fix warnings

+++ Modules/sha3module.c
@@ -502,7 +502,8 @@ _SHAKE_digest(SHA3object *self, Py_ssize_t digestlen, int hex)
      * - the output length is zero -- we follow the existing behavior and return
      *   an empty digest, without raising an error */
     if (digestlen > 0) {
-        (void)Hacl_Hash_SHA3_squeeze(self->hash_state, digest, digestlen);
+        (void)Hacl_Hash_SHA3_squeeze(self->hash_state, digest,
+                                     (uint32_t)digestlen);
     }
     if (hex) {
         result = _Py_strhex((const char *)digest, digestlen);
